KernelInterpolation.jl API

KernelInterpolation.AbstractKernel — Type

AbstractKernel

An abstract supertype of kernels.

KernelInterpolation.GaussKernel — Type

GaussKernel{Dim}(; shape_parameter = 1.0)

Gaussian kernel function with

\[ \phi(r) = \exp(-(\varepsilon r)^2),\]

where $\varepsilon$ is the shape parameter. The Gaussian kernel is always positive definite. See Wendland (2004), p. 74.

See also RadialSymmetricKernel.

Holger Wendland (2004) Scattered Data Approximation Cambridge University Press DOI: 10.1017/CBO9780511617539

KernelInterpolation.Interpolation — Type

Interpolation

Interpolation object that can be evaluated at a node and represents a kernel interpolation of the form

\[ s(x) = \sum_{j = 1}^n c_jK(x, x_j) + \sum_{k = 1}^q d_kp_k(x),\]

where $x_j$ are the nodes in the nodeset and $s(x)$ the interpolant satisfying $s(x_j) = f(x_j)$, where $f(x_j)$ are given by values in interpolate and $p_k$ is a basis of the q-dimensional space of multivariate polynomials of order order. The additional conditions

\[ \sum_{j = 1}^n c_jp_k(x_j) = 0, \quad k = 1,\ldots, q\]

are enforced.

KernelInterpolation.InverseMultiquadricKernel — Type

InverseMultiquadricKernel{Dim}(beta = 0.5; shape_parameter = 1.0)

Inverse multiquadric kernel function with

\[ \phi(r) = (1 + (\varepsilon r)^2)^{-\beta},\]

where $\varepsilon$ is the shape parameter. The inverse multiquadric kernel is always positive definite. See Wendland (2004), p. 76 and p. 95.

See also RadialSymmetricKernel.

Holger Wendland (2004) Scattered Data Approximation Cambridge University Press DOI: 10.1017/CBO9780511617539

KernelInterpolation.Matern12Kernel — Type

Matern12Kernel{Dim}(; shape_parameter = 1.0)

Matern kernel with $\nu = 1/2$, i.e.

\[ \phi(r) = \exp(-\varepsilon r),\]

where $\varepsilon$ is the shape parameter. The Matern kernel is positive definite.

See Wikipedia.

See also MaternKernel, RadialSymmetricKernel.

KernelInterpolation.Matern32Kernel — Type

Matern32Kernel{Dim}(; shape_parameter = 1.0)

Matern kernel with $\nu = 3/2$, i.e.

\[ \phi(r) = (1 + \sqrt{3}\varepsilon r)\exp(-\sqrt{3}\varepsilon r),\]

where $\varepsilon$ is the shape parameter. The Matern kernel is positive definite.

See Wikipedia.

See also MaternKernel, RadialSymmetricKernel.

KernelInterpolation.Matern52Kernel — Type

Matern52Kernel{Dim}(; shape_parameter = 1.0)

Matern kernel with $\nu = 5/2$, i.e.

\[ \phi(r) = (1 + \sqrt{5}\varepsilon r + 5\cdot(\varepsilon r)^2/3)\exp(-\sqrt{5}\varepsilon r),\]

where $\varepsilon$ is the shape parameter. The Matern kernel is positive definite.

See Wikipedia.

See also MaternKernel, RadialSymmetricKernel.

KernelInterpolation.Matern72Kernel — Type

Matern72Kernel{Dim}(; shape_parameter = 1.0)

Matern kernel with $\nu = 7/2$, i.e.

\[ \phi(r) = (1 + \sqrt{7}\varepsilon r + 12\cdot(\varepsilon r)^2/5 + 7\cdot(\varepsilon r)^3/15)\exp(-\sqrt{7}\varepsilon r),\]

where $\varepsilon$ is the shape parameter. The Matern kernel is positive definite.

See Wikipedia.

See also MaternKernel, RadialSymmetricKernel.

KernelInterpolation.MaternKernel — Type

MaternKernel{Dim}(nu = 1.5; shape_parameter = 1.0)

Matern kernel with

\[ \phi(r) = \frac{2^{1-\nu}}{\Gamma(\nu)}\big(\sqrt{2\nu}\varepsilon r\big)^\nu K_\nu\big(\sqrt{2\nu}\varepsilon r\big),\]

where $\varepsilon$ is the shape parameter. The Matern kernel is positive definite.

See Wikipedia.

See also RadialSymmetricKernel.

KernelInterpolation.MultiquadricKernel — Type

MultiquadricKernel{Dim}(beta = 0.5; shape_parameter = 1.0)

Multiquadric kernel function with

\[ \phi(r) = (1 + (\varepsilon r)^2)^\beta,\]

where $\varepsilon$ is the shape parameter. The multiquadric kernel is conditionally positive definite of order $m = \lceil\beta \rceil$. See Wendland (2004), p. 109.

See also RadialSymmetricKernel.

Holger Wendland (2004) Scattered Data Approximation Cambridge University Press DOI: 10.1017/CBO9780511617539

KernelInterpolation.NodeSet — Type

NodeSet(nodes)

Set of interpolation nodes.

KernelInterpolation.PolyharmonicSplineKernel — Type

PolyharmonicSplineKernel{Dim}(k)

Polyharmonic spline kernel function with

\[ \phi(r) = \begin{cases} r^k, &\text{ if } k \text{ odd}\\ r^k\log(r), &\text{ if } k \text{ even} \end{cases}.\]

The polyharmonic spline is conditionally positive definite of order $m = \lceil k/2\rceil$ for odd k and order $m = k/2 + 1$ for even k. See Wendland (2004), pp. 111–112.

See also RadialSymmetricKernel.

Holger Wendland (2004) Scattered Data Approximation Cambridge University Press DOI: 10.1017/CBO9780511617539

KernelInterpolation.RadialCharacteristicKernel — Type

RadialCharacteristicKernel{Dim}(beta = 2.0; shape_parameter = 1.0)

Radial characteristic function kernel function with

\[ \phi(r) = (1 - (\varepsilon r)^2)^\beta_+,\]

where $\varepsilon$ is the shape parameter. The radial characteristic function is positive definite if $\beta\ge (d + 1)/2$. It is compactly supported. See Wendland (2004), p. 80.

See also RadialSymmetricKernel.

Holger Wendland (2004) Scattered Data Approximation Cambridge University Press DOI: 10.1017/CBO9780511617539

KernelInterpolation.RadialSymmetricKernel — Type

RadialSymmetricKernel

An abstract supertype of radial symmetric kernels. Radial symmetric kernels are generated by an even and continuous function $\Phi: \mathbb{R}^d\to\mathbb{R}$, which is radial-symmetric meaning that there exists a $\phi:\mathbb{R}\to\mathbb{R}$ such that

\[ \Phi(x) = \phi(\Vert x\Vert).\]

The kernel is then defined by

\[ K(x, y) = \Phi(x - y).\]

KernelInterpolation.RieszKernel — Type

RieszKernel{Dim}(beta; shape_parameter = 1.0)

Riesz kernel with

\[ \phi(r) = -(\varepsilon r)^\beta,\]

where $\varepsilon$ is the shape parameter and $\beta\in (0,2)$. The Riesz kernel is conditionally positive definite of order 1. See Hertrich et al. (2023).

See also RadialSymmetricKernel.

Johannes Hertrich, Christian Wald, Fabian Altekrüger, Paul Hagemann (2023) Generative Sliced MMD Flows with Riesz Kernels ArXiv: 2305.11463

KernelInterpolation.ThinPlateSplineKernel — Type

ThinPlateSplineKernel{Dim}()

Thin plate spline kernel function with

\[ \phi(r) = r^2\log(r),\]

i.e. PolyharmonicSplineKernel with $k = 2$. The thin plate spline is conditionally positive definite of order $m = 2$. See Wendland (2004), p. 112.

See also RadialSymmetricKernel.

Holger Wendland (2004) Scattered Data Approximation Cambridge University Press DOI: 10.1017/CBO9780511617539

KernelInterpolation.WendlandKernel — Type

WendlandKernel{Dim}(k; shape_parameter = 1.0, d = Dim)

Wendland kernel with

\[ \phi(r) = \begin{cases} p(\varepsilon r), \text{ if } 0\le \varepsilon r\le 1\\ 0, \text{ if } \varepsilon r > 1 \end{cases},\]

where $\varepsilon$ is the shape parameter and $p$ is a polynomial. The Wendland kernel is positive definite and compactly supported. See Wendland (2004), p. 129.

See also RadialSymmetricKernel.

Holger Wendland (2004) Scattered Data Approximation Cambridge University Press DOI: 10.1017/CBO9780511617539

KernelInterpolation.coefficients — Method

coefficients(itp::Interpolation)

Obtain all the coefficients of the linear combination for the interpolant, i.e. both the coefficients for the kernel part and for the polynomial part.

See also kernel_coefficients and polynomial_coefficients.

KernelInterpolation.default_example — Method

default_example()

Return the path to an example that can be used to quickly see KernelInterpolation.jl in action. See also examples_dir and get_examples.

KernelInterpolation.dim — Method

dim(kernel)

Return the dimension of a kernel, i.e. the size of the input vector.

KernelInterpolation.dim — Method

dim(itp::Interpolation)

Return the dimension of the input variables of the interpolation.

KernelInterpolation.examples_dir — Method

examples_dir()

Return the directory where the example files provided with KernelInterpolation.jl are located.

Examples

readdir(examples_dir())

KernelInterpolation.get_examples — Method

get_examples()

Return a list of all examples that are provided by KernelInterpolation.jl. See also examples_dir and default_example.

KernelInterpolation.get_name — Method

get_name(kernel::AbstractKernel)

Returns the canonical, human-readable name for the given system of equations.

KernelInterpolation.homogeneous_hypercube — Function

homogeneous_hypercube(n, x_min = ntuple(_ -> 0.0, dim), x_max = ntuple(_ -> 1.0, dim); [dim])

If n is integer, create a NodeSet with n homogeneously distributed nodes in every dimension each of dimension dim inside a hypercube defined by the bounds x_min and x_max. If n is a Tuple of length dim, then use as many nodes in each dimension as described by n. The resulting NodeSet will have $n^{\textrm{dim}}$ respectively $\prod_{j = 1}n_j$ points. If the bounds are given as single values, they are applied for each dimension. If they are Tuples of size dim, the hypercube has the according bounds. If dim is not given explicitly, it is inferred by the lengths of n, x_min and x_max if possible.

KernelInterpolation.homogeneous_hypercube_boundary — Method

homogeneous_hypercube_boundary(n, x_min = ntuple(_ -> 0.0, dim), x_max = ntuple(_ -> 1.0, dim); [dim])

If n is integer, create a NodeSet with n homogeneously distributed nodes in every dimension each of dimension dim on the boundary of a hypercube defined by the bounds x_min and x_max. If n is a Tuple of length dim, then use as many nodes in each dimension as described by n. If the bounds are given as single values, they are applied for each dimension. If they are Tuples of size dim, the hypercube has the according bounds. If dim is not given explicitly, it is inferred by the lengths of n, x_min and x_max if possible.

KernelInterpolation.include_example — Method

include_example([mod::Module=Main,] example::AbstractString; kwargs...)

include the file example and evaluate its content in the global scope of module mod. You can override specific assignments in example by supplying keyword arguments. It's basic purpose is to make it easier to modify some parameters while running KernelInterpolation from the REPL. Additionally, this is used in tests to reduce the computational burden for CI while still providing examples with sensible default values for users.

Copied from Trixi.jl.

KernelInterpolation.interpolate — Method

interpolate(nodeset, values, kernel = GaussKernel{dim(nodeset)}(), m = order(kernel))

Interpolate the values evaluated at the nodes in the nodeset to a function using the kernel kernel and polynomials up to a degree polynomial_degree, i.e. determine the coefficients c_j and d_k in the expansion

\[ s(x) = \sum_{j = 1}^n c_jK(x, x_j) + \sum_{k = 1}^q d_kp_k(x),\]

where $x_j$ are the nodes in the nodeset and $s(x)$ the interpolant $s(x_j) = f(x_j)$, where $f(x_j)$ are given by values and $p_k$ is a basis of the q-dimensional space of multivariate polynomials with maximum degree of m - 1. If m = 0, no polynomial is added. The additional conditions

\[ \sum_{j = 1}^n c_jp_k(x_j) = 0, \quad k = 1,\ldots, q\]

are enforced. Returns an Interpolation object.

KernelInterpolation.interpolation_kernel — Method

interpolation_kernel(itp)

Return the kernel from an interpolation object.

KernelInterpolation.kernel_coefficients — Method

kernel_coefficients(itp::Interpolation)

Obtain the coefficients of the kernel part of the linear combination for the interpolant.

See also coefficients and polynomial_coefficients.

KernelInterpolation.kernel_inner_product — Method

kernel_inner_product(itp1, itp2)

Inner product of the native space for two interpolants itp1 and itp2 with the same kernel. The inner product is defined as

\[ (f, g)_K = \sum_{j = 1}^n\sum_{k = 1}^mc_jd_kK(x_j, y_k)\]

for the interpolants $f(x) = \sum_{j = 1}^nc_jK(x, x_j)$ and $g(x) = \sum_{k = 1}^md_kK(x, y_k)$.

See also kernel_norm.

KernelInterpolation.kernel_norm — Method

kernel_norm(itp)

Norm of the native space defined by the kernel of the interpolant itp. The norm is defined as

\[ \|f\|_K^2 = \sum_{j,k=1}^nc_jc_kK(x_j, x_k)\]

for the interpolant $f(x) = \sum_{j = 1}^nc_jK(x, x_j)$.

See also kernel_inner_product.

KernelInterpolation.nodeset — Method

nodeset(itp)

Return the node set from an interpolation object.

KernelInterpolation.order — Method

order(itp)

Return the order $m$ of the polynomial used for the interpolation, i.e. the polynomial degree plus 1. If $m = 0$, no polynomial is added.

KernelInterpolation.polynomial_basis — Method

polynomial_basis(itp::Interpolation)

Return a vector of the polynomial basis functions used for the interpolation.

See also polyvars.

KernelInterpolation.polynomial_coefficients — Method

polynomial_coefficients(itp::Interpolation)

Obtain the coefficients of the polynomial part of the linear combination for the interpolant.

See also coefficients and kernel_coefficients.

KernelInterpolation.polyvars — Method

polyvars(itp::Interpolation)

Return a vector of the polynomial variables.

See also polynomial_basis.

KernelInterpolation.random_hypercube — Function

random_hypercube(n, x_min = ntuple(_ -> 0.0, dim), x_max = ntuple(_ -> 1.0, dim); [dim])

Create a NodeSet with n random nodes each of dimension dim inside a hypercube defined by the bounds x_min and x_max. If the bounds are given as single values, they are applied for each dimension. If they are Tuples of size dim the hypercube has the according bounds. If dim is not given explicitly, it is inferred by the lengths of x_min and x_max if possible.

KernelInterpolation.random_hypercube_boundary — Function

random_hypercube_boundary(n, x_min = ntuple(_ -> 0.0, dim), x_max = ntuple(_ -> 1.0, dim); [dim])

Create a NodeSet with n random nodes each of dimension dim on the boundary of a hypercube defined by the bounds x_min and x_max. If the bounds are given as single values, they are applied for each dimension. If they are Tuples of size dim the hypercube has the according bounds. If dim is not given explicitly, it is inferred by the lengths of x_min and x_max if possible.

KernelInterpolation.random_hypersphere — Function

random_hypersphere(n, r = 1.0, center = zeros(dim); [dim])

Create a NodeSet with n random nodes each of dimension dim inside a hypersphere with radius r around the center center. If dim is not given explicitly, it is inferred by the length of center if possible.

KernelInterpolation.random_hypersphere_boundary — Function

random_hypersphere_boundary(n, r = 1.0, center = zeros(dim); [dim])

Create a NodeSet with n random nodes each of dimension dim at the boundary of a hypersphere with radius r around the center center. If dim is not given explicitly, it is inferred by the length of center if possible.

KernelInterpolation.separation_distance — Method

separation_distance(nodeset::NodeSet)

Return the separation distance of a node set $X = \{x_1,\ldots, x_n\}$ defined by

\[ q_X = \frac{1}{2}\min_{x_i\neq x_j}\|x_i - x_j\|.\]

KernelInterpolation.system_matrix — Method

system_matrix(itp::Interpolation)

Return the system matrix, i.e. the matrix

\[\begin{pmatrix} A & P \\ P^T & 0 \end{pmatrix},\]

where $A\in\mathbb{R}^{n\times n}$ is the matrix with entries $a_{ij} = K(x_i, x_j)$ for the kernel function K and nodes x_i and $P\in\mathbb{R}^{n\times q}$ is the matrix with entries $p_{ij} = p_j(x_i)$, where $p_j$ is the $j$-th multivariate monomial of the space of polynomials up to degree $m$.

KernelInterpolation.values_along_dim — Method

values_along_dim(nodeset::NodeSet, i::Int)

Convenience function to return all $x_i$-values of the nodes, i.e. the i-th component of each node. Supported for nodeset with dim(nodeset) >= i.